Revert "SAM-47,  Изменил логику применения изменений."
This reverts commit 415ccadd6e0717e1b7c422a82d7ee0297b2afed7.
