--- PhotoTune/Services/ImageProcessing/ChangedInstrument.swift
@@ -1,21 +1,0 @@
-//
-//  ChangedInstrument.swift
-//  PhotoTune
-//
-//  Created by Mikhail Medvedev on 14.01.2020.
-//  Copyright © 2020 Mikhail Medvedev. All rights reserved.
-//
-
-import Foundation
-
-enum ChangedInstrument
-{
-	case autoEnhance
-	case filter
-	case colorControls
-	case vignette
-	case sharpness
-	case rotation
-	case all
-	case none
-}
